+++ tests/test_model.py
@@ -1,0 +1,8 @@
+from scripts.model import get_Model
+
+
+def test_model():
+    model = get_Model()
+    
+    
+